% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html.R
\name{html_tweets}
\alias{html_tweets}
\title{Render tweets}
\usage{
html_tweets(x, title = NULL)
}
\arguments{
\item{x}{Timeline data returned by \link[rtweet]{get_my_timeline}.}

\item{title}{Optional, title of rendered tweets document.}
}
\value{
Saves a temporary HTML file.
}
\description{
Render non-retweets as temporary HTML file
}
