% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/favorites.R
\name{view_favorites}
\alias{view_favorites}
\title{Render favorited tweets}
\usage{
view_favorites(user = NULL)
}
\arguments{
\item{user}{Name of user; used to determine favorites. If NULL (default),
this returns the favorites of the authenticating user. If a user name is
provided, the displayed tweets will have been favorited by that user.}
}
\value{
Favorited tweets will be displayed as temporary HTML file.
}
\description{
Render tweets favorited by a given user
}
