% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{view_search}
\alias{view_search}
\title{Render Twitter search results}
\usage{
view_search(query)
}
\arguments{
\item{query}{Search query used to find tweets.}
}
\value{
Matching tweets will be displayed as temporary HTML file.
}
\description{
Render tweets matched to Twitter search
}
