% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeline.R
\name{view_timeline}
\alias{view_timeline}
\title{Render timeline tweets}
\usage{
view_timeline(user = NULL)
}
\arguments{
\item{user}{Name of user; used to determine timeline. If NULL (default), this
returns the home timeline of the authenticating user. If a user name is
provided, the timeline will display tweets sent by that user.}
}
\value{
Timeline will be displayed as temporary HTML file.
}
\description{
Render tweets posted to home (authenticating) user's timeline
}
